requirements.txt
lightning==2.5.0.post0
matplotlib==3.10.3
numpy==2.3.1
optuna==4.2.1
pandas==2.3.0
Pillow==11.2.1
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchinfo==1.8.0
torchmetrics==1.6.1
torchvision==0.20.1
tqdm==4.67.1
